+++ apps/nbs-gateway/src/main/java/gov/cdc/nbs/gateway/welcome/WelcomeService.java
@@ -1,0 +1,12 @@
+package gov.cdc.nbs.gateway.welcome;
+
+import gov.cdc.nbs.gateway.Service;
+
+import java.net.URI;
+
+public class WelcomeService extends Service {
+
+  public WelcomeService(final URI uri, final String base) {
+    super(uri, base);
+  }
+}
